Added batch.txt
to run, simply have this file in the root user directory on your ssh terminal (i.e /users/PAS2622/mdsil11) and run 'sbatch batch.txt' or modify the relative path of the python3 command. You can track the status of the job on the jobs dashboard on osc. The log-file should also populate in the same directory and update every 5-10 minutes.